requirements.txt
httpx==0.27.0
motor==3.7.0
pymongo==4.11.3
urllib3==2.2.2
openai==1.42.0
pillow==10.4.0
pyyaml==6.0.1
pydantic-core==2.33.1
pydantic==2.11.2
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.3
yarl==1.19.0
zipp==3.19.1
python-json-logger==2.0.7
fastapi==0.115.12
starlette==0.46.1
uvicorn==0.30.3
jsonschema==4.22.0
genson==1.3.0
datamodel-code-generator==0.25.6
sentry-sdk==2.25.1
cryptography==43.0.0
google-auth==2.35.0
kombu==5.3.7
taskiq==0.11.16
taskiq-redis==1.0.4
certifi==2025.1.31
hiredis==2.3.2
boto3==1.34.149
workflowai==0.6.4
soundfile==0.12.1
pydub==0.25.1
azure-storage-blob==12.21.0
python-multipart==0.0.18
pdf2image==1.17.0
stripe==11.4.1
html2text==2024.2.26
jinja2==3.1.5
clickhouse-connect==0.8.16
python-dotenv==1.1.0
aiohttp==3.11.16
